Cambios

+++ src/controllers/productController.js
@@ -3,11 +3,14 @@ const { Product } = require("../models/Product.js");
 const productController = {
     createProduct: async (req, res) => {
         try {
+            if (!req.body?.user) {
+                return res.status(403).json({message: "Action restricted"});
+            }
             const product = await Product.create(req.body);
             res.status(201).json(product);
         } catch (error) {
-            console.log(error);
-            res.status(501).send({message: "There was a problem trying to create the product"});
+            console.error(error);
+            res.status(501).json({message: "There was a problem trying to create the product"});
         }
     },
     getProducts: async (req, res) => {
@@ -15,8 +18,8 @@ const productController = {
             const products = await Product.find();
             res.status(201).json(products);
         } catch (error) {
-            console.log(error);
-            res.status(501).send({message: "There was a problem trying to get all products"});
+            console.error(error);
+            res.status(501).json({message: "There was a problem trying to get all products"});
         }
     },
     getProductById: async (req, res) => {
@@ -24,42 +27,44 @@ const productController = {
         try {
             const product = await Product.findById(id);
             if (!product) {
-                res.status(404).send({message: "There is no product with that id"});
+                res.status(404).json({message: "There is no product with that id"});
             }
-            res.status(201).send(product);
+            res.status(201).json(product);
         } catch (error) {
-            console.log(error);
-            res.status(501).send({message: "There was a problem trying to get the product"});
+            console.error(error);
+            res.status(501).json({message: "There was a problem trying to get the product"});
         }
     },
     updateProduct: async (req, res) => {
         const id = req.params.id;
-        console.log("=== UPDATE PRODUCT ===");
-        console.log("ID:", id);
-        console.log("Method:", req.method);
-        console.log("Body:", req.body);
         try {
-            const task = await Product.findByIdAndUpdate(id, req.body);
-            if (!task) {
-                return res.status(404).send({message: "There is no product with that id"});
+            if (!req.body?.user) {
+                return res.status(403).json({message: "Action restricted"});
+            }
+            const product = await Product.findByIdAndUpdate(id, req.body);
+            if (!product) {
+                return res.status(404).json({message: "There is no product with that id"});
             }
         } catch (error) {
-            console.log(error);
-            res.status(501).send({message: "There was a problem trying to update the product"});
+            console.error(error);
+            res.status(501).json({message: "There was a problem trying to update the product"});
         }
     },
     deleteProduct: async (req, res) => {
         const id = req.params.id;
         try {
+            if (!req.params?.user) {
+                return res.status(403).json({message: "Action restricted"});
+            }
             const product = await Product.findById(id);
             if (!product) {
                 res.status(404).send({message: "There is no product with that id"});
             }
             const deleteCount = await Product.deleteOne({_id: id});
-            res.status(201).send({message: "Product successfully deleted", deleteCount: deleteCount});
+            res.status(201).json({message: "Product successfully deleted", deleteCount: deleteCount});
         } catch (error) {
-            console.log(error);
-            res.status(501).send({message: "There was a problem trying to delete the product"});
+            console.error(error);
+            res.status(501).json({message: "There was a problem trying to delete the product"});
         }
     },
 };
